requirements.txt
matrix-nio
matrix-nio[e2e]
dotenv
